demo/backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7

torch>=2.1
transformers>=4.40
scikit-learn>=1.3
numpy>=1.26

google-api-python-client>=2.140
google-auth-httplib2>=0.2
google-auth-oauthlib>=1.2
